Dev: move chart components and scss in site folder to charts folder
