+++ charts/controls/CommandPalette.scss
