+++ charts/controls/Controls.tsx
@@ -9,7 +9,7 @@ import { ChartConfig } from "charts/core/ChartConfig"
 import { getQueryParams, getWindowQueryParams } from "utils/client/url"
 import { ChartView } from "charts/core/ChartView"
 import { Timeline } from "./Timeline"
-import { extend, keys, entries, max, formatValue } from "../utils/Util"
+import { extend, keys, entries, max, formatValue } from "charts/utils/Util"
 import { ADMIN_BASE_URL, ENV } from "settings"
 
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
@@ -26,7 +26,7 @@ import { faExchangeAlt } from "@fortawesome/free-solid-svg-icons/faExchangeAlt"
 import { faTwitter } from "@fortawesome/free-brands-svg-icons/faTwitter"
 import { faFacebook } from "@fortawesome/free-brands-svg-icons/faFacebook"
 import { faExternalLinkAlt } from "@fortawesome/free-solid-svg-icons/faExternalLinkAlt"
-import { TimeBound } from "../utils/TimeBounds"
+import { TimeBound } from "charts/utils/TimeBounds"
 import { HighlightToggleConfig } from "charts/core/ChartConstants"
 
 @observer
