+++ charts/controls/CountryPicker.scss
