+++ charts/loadingIndicator/LoadingIndicator.scss
