+++ charts/mapCharts/MapTab.tsx
@@ -20,7 +20,7 @@ import { select } from "d3-selection"
 import { easeCubic } from "d3-ease"
 import { ChartLayout, ChartLayoutView } from "charts/core/ChartLayout"
 import { ChartView } from "charts/core/ChartView"
-import { LoadingOverlay } from "charts/core/LoadingOverlay"
+import { LoadingOverlay } from "charts/loadingIndicator/LoadingOverlay"
 import { ControlsOverlay } from "charts/controls/ControlsOverlay"
 import { MapTooltip } from "./MapTooltip"
 import { ProjectionChooser } from "./ProjectionChooser"
