+++ charts/mapCharts/ProjectionChooser.tsx
@@ -3,7 +3,7 @@ import { computed, action } from "mobx"
 import { observer } from "mobx-react"
 import Select, { ValueType } from "react-select"
 import { worldRegions, labelsByRegion } from "./WorldRegions"
-import { first } from "../utils/Util"
+import { first } from "charts/utils/Util"
 import { Bounds } from "charts/utils/Bounds"
 import { MapProjection } from "./MapProjections"
 import { asArray, getStylesForTargetHeight } from "utils/client/react-select"
