+++ charts/scatterCharts/MultiColorPolyline.tsx
@@ -2,7 +2,7 @@ import * as React from "react"
 import { computed } from "mobx"
 import { observer } from "mobx-react"
 
-import { last } from "../utils/Util"
+import { last } from "charts/utils/Util"
 
 interface MultiColorPolylinePoint {
     x: number
