+++ charts/sparkBars/SparkBars.scss
