+++ explorer/client/ExplorerShell.scss
