+++ site/client/SearchCountry.tsx
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom"
 import Select, { ValueType } from "react-select"
 import { countries } from "utils/countries"
 import { asArray } from "utils/client/react-select"
-import { Analytics } from "./Analytics"
+import { Analytics } from "charts/core/Analytics"
 import { sortBy } from "charts/utils/Util"
 import { countryProfileSpecs } from "../server/countryProfileProjects"
